feat: 实现 Agent/Session/Knowledge 核心模块 (Phase 2)
主要变更:
- 新增 internal/apiserver/model/*.gen.go 数据模型
- 新增 internal/apiserver/store/ 数据访问层
- 新增 internal/apiserver/biz/v1/{agent,session,knowledge}/ 业务逻辑
- 新增 internal/apiserver/handler/http/{agent,session,knowledge}.go HTTP 处理器
- 新增 internal/pkg/agent/ Agent 抽象层（ChatModel/ReactAgent/Tool）
- 新增 internal/pkg/token/ Token 服务
- 新增 pkg/api/apiserver/v1/{agent,session,knowledge}.proto API 定义
- 移除 Post 示例模块及相关代码
- 集成 Eino 依赖（cloudwego/eino）
- 清理第三方 protobuf 文件

+++ examples/client/user/main.go
@@ -136,7 +136,7 @@ func main() {
 	}
 	log.Printf("[GetUser        ] Success to get user: %v", createUserResponse.UserID)
 
-	listResponse, err := client.ListUser(ctx, &apiv1.ListUserRequest{Offset: 0, Limit: *limit})
+	listResponse, err := client.ListUser(ctx, &apiv1.ListUserRequest{Offset: 0, Limit: uint64(*limit)})
 	if err != nil {
 		if !strings.Contains(err.Error(), "PermissionDenied") {
 			log.Printf("Failed to list user: %v", err)
@@ -151,7 +151,7 @@ func main() {
 	ctx = helper.MustWithAdminToken(ctx, client)
 
 	// 请求 ListUser 接口
-	listResponse, err = client.ListUser(ctx, &apiv1.ListUserRequest{Offset: 0, Limit: *limit})
+	listResponse, err = client.ListUser(ctx, &apiv1.ListUserRequest{Offset: 0, Limit: uint64(*limit)})
 	if err != nil {
 		log.Printf("Failed to list user: %v", err)
 		return
